# api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Kazer Universal API",
    "version": "0.1.0",
    "description": "Discovery API for the public Kazer Universal MVP. Capability execution is disabled until a provider is configured."
  },
  "paths": {
...
    "/v1/adapters/{id}/chat": { "post": { "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "string", "enum": ["llama-cpp", "qwen3", "deepseek-r1"] } }], "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["messages"], "properties": { "model": { "type": "string" }, "messages": { "type": "array" }, "max_tokens": { "type": "integer", "maximum": 8192 }, "temperature": { "type": "number" }, "stream": { "type": "boolean" } } } } } }, "responses": { "200": { "description": "Upstream OpenAI-compatible response" }, "401": { "description": "Kazer API key required when database is active" }, "501": { "description": "Adapter not configured" }, "502": { "description": "Upstream adapter error" } } } },
    "/v1/repository/inspect": { "post": { "summary": "Read public GitHub repository metadata safely", "description": "Reads public README, license files and text manifests without cloning or executing external code. Only HTTPS github.com repository URLs are accepted.", "requestBody": { "required": true, "content": { "application/json": { "schema": { "type": "object", "required": ["url"], "properties": { "url": { "type": "string", "format": "uri" } } } } } }, "responses": { "200": { "description": "Inspection result with license, suggested skills, capabilities and evidence" }, "400": { "description": "Invalid or disallowed repository URL" }, "502": { "description": "Public metadata could not be read" } } } },
...
  }
}